requirements.txt
aiogram==3.21.0
aiosqlite==0.20.0
python-dotenv